refactor(cli): move hatchLocal and retireLocal to lib to fix cross-command imports
